+++ CHANGELOG.md
@@ -1,7 +1,11 @@
 # Changelog - emilohman.nu
 ## 1.0.9 - 
+**Added:**
+- Added subtitle in [Header.jsx](src/components/Header/Header.jsx)
+
 **Changed:**
 - Changed Websites to Work
+- Made changes to the start animation
 
 ## 1.0.8 - _2019-08-18_
 **Added:**